[FEAT]: FeedHistory Coordinator 구현
